Move TracepointServiceImpl to its own module
Part of a series of changes that aim to isolate each gRPC service into its own
module for better readability and maintenance. Currently, the Service module
contains multiple independent subsets of classes contributing to the different
services, so we extract those.

Test: Run Orbit with `--enable_tracepoint_feature` and verify list of
tracepoints in the "Tracepoints" tab.
