+++ CMakeLists.txt
@@ -149,6 +149,7 @@ else()
   add_subdirectory(src/OrbitVulkanLayer)
   add_subdirectory(src/ProducerSideChannel)
   add_subdirectory(src/Service)
+  add_subdirectory(src/TracepointService)
   add_subdirectory(src/UserSpaceInstrumentation)
   add_subdirectory(src/VulkanTutorial)
 endif()
