+++ src/Service/ServiceUtilsTest.cpp
@@ -138,67 +138,4 @@ TEST(ServiceUtils, FindSymbolsFilePath) {
   }
 }
 
-TEST(ServiceUtils, CategoriesTracepoints) {
-  using orbit_grpc_protos::TracepointInfo;
-
-  if (getuid() != 0) {
-    GTEST_SKIP() << "test is supported with root uid";
-  }
-
-  const auto tracepoint_infos = utils::ReadTracepoints();
-
-  std::deque<std::string> categories;
-  std::transform(tracepoint_infos.value().cbegin(), tracepoint_infos.value().cend(),
-                 std::back_inserter(categories),
-                 [](const TracepointInfo& value) { return value.category(); });
-
-  ASSERT_GT(categories.size(), 0);
-  static const std::array<std::string, 10> kCategoriesAvailable = {
-      "sched",    "task",    "module",       "signal",     "sock",
-      "syscalls", "migrate", "raw_syscalls", "exceptions", "iomap"};
-
-  static const std::array<std::string, 3> kCategoriesUnavailable = {"orbit", "profiler",
-                                                                    "instrumentation"};
-
-  for (std::string category_available : kCategoriesAvailable) {
-    ASSERT_TRUE(find(categories.begin(), categories.end(), category_available) != categories.end());
-  }
-
-  for (std::string category_unavailable : kCategoriesUnavailable) {
-    ASSERT_TRUE(find(categories.begin(), categories.end(), category_unavailable) ==
-                categories.end());
-  }
-}
-
-TEST(ServiceUtils, NamesTracepoints) {
-  using orbit_grpc_protos::TracepointInfo;
-
-  if (getuid() != 0) {
-    GTEST_SKIP() << "test is supported with root uid";
-  }
-
-  const auto tracepoint_infos = utils::ReadTracepoints();
-
-  std::deque<std::string> names;
-  std::transform(tracepoint_infos.value().cbegin(), tracepoint_infos.value().cend(),
-                 std::back_inserter(names),
-                 [](const TracepointInfo& value) { return value.name(); });
-
-  ASSERT_TRUE(names.size() > 0);
-  static const std::array<std::string, 10> kNamesAvailable = {
-      "sched_switch", "sched_wakeup",    "sched_process_fork", "sched_waking", "task_rename",
-      "task_newtask", "signal_generate", "signal_deliver",     "timer_init",   "timer_start"};
-
-  static const std::array<std::string, 5> kNamesUnavailable = {
-      "orbit", "profiler", "instrumentation", "enable", "filter"};
-
-  for (std::string name_available : kNamesAvailable) {
-    ASSERT_TRUE(find(names.begin(), names.end(), name_available) != names.end());
-  }
-
-  for (std::string name_unavailable : kNamesUnavailable) {
-    ASSERT_TRUE(find(names.begin(), names.end(), name_unavailable) == names.end());
-  }
-}
-
 }  // namespace orbit_service::utils
