+++ src/TracepointService/include/TracepointService/TracepointServiceImpl.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef ORBIT_TRACEPOINTSERVICEIMPL_H
-#define ORBIT_TRACEPOINTSERVICEIMPL_H
+#ifndef TRACEPOINT_SERVICE_TRACEPOINT_SERVICE_IMPL_
+#define TRACEPOINT_SERVICE_TRACEPOINT_SERVICE_IMPL_
 
 #include <grpcpp/grpcpp.h>
 
@@ -13,7 +13,7 @@
 #include "services.grpc.pb.h"
 #include "services.pb.h"
 
-namespace orbit_service {
+namespace orbit_tracepoint_service {
 
 using orbit_grpc_protos::GetTracepointListRequest;
 using orbit_grpc_protos::GetTracepointListResponse;
@@ -26,5 +26,6 @@ class TracepointServiceImpl final : public TracepointService::Service {
                                                GetTracepointListResponse* response) override;
 };
 
-}  // namespace orbit_service
-#endif  // ORBIT_TRACEPOINTSERVICEIMPL_H
+}  // namespace orbit_tracepoint_service
+
+#endif  // TRACEPOINT_SERVICE_TRACEPOINT_SERVICE_IMPL_
